% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param_cobalt}
\alias{param_cobalt}
\title{Construct parameter list for cobalt}
\usage{
param_cobalt(
  stats = c("m", "v", "ks", "ovl"),
  continuous = "std",
  binary = "std",
  abs = TRUE,
  s.d.denom = "pooled"
)
}
\arguments{
\item{stats}{character; which statistic(s) should be reported.}

\item{continuous}{whether mean differences for continuous variables
should be standardized ("std") or raw ("raw").}

\item{binary}{whether mean differences for binary variables (i.e.,
difference in proportion) should be standardized ("std") or raw ("raw").}

\item{abs}{logical; whether displayed balance statistics
should be in absolute value or not.}

\item{s.d.denom}{character; how the denominator for standardized
mean differences should be calculated, if requested.}
}
\value{
A \code{list} of parameters that can be passed to \code{\link[=missDiag]{missDiag()}}.
}
\description{
Construct parameter list for cobalt
}
\details{
For more information about these parameters, see
\code{\link[cobalt:bal.tab]{cobalt::bal.tab()}}.
}
