% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param_sbw}
\alias{param_sbw}
\title{Construct parameter list for sbw}
\usage{
param_sbw(
  bal_alg = TRUE,
  bal_tol = 0,
  bal_std = "group",
  bal_gri = c(0.0001, 0.001, 0.002, 0.005, 0.01, 0.02, 0.05, 0.1),
  bal_sam = 1000,
  sol_nam = "quadprog"
)
}
\arguments{
\item{bal_alg}{use tuning algorithm as in Wang and Zubizarreta (2020)?}

\item{bal_tol}{tolerance level.}

\item{bal_std}{tolerances adjustment.}

\item{bal_gri}{grid of values for the tuning algorithm.}

\item{bal_sam}{number of replicates to be used by the tuning algorithm.}

\item{sol_nam}{solver name. Either "cplex", "gurobi", "mosek",
"pogs", or "quadprog".}
}
\value{
A \code{list} of parameters that can be passed to \code{\link[=missDiag]{missDiag()}}.
}
\description{
Construct parameter list for sbw
}
\details{
For more information about these parameters, see \code{\link[sbw:sbw]{sbw::sbw()}}.
}
